+++ src/class-main.php
@@ -2,6 +2,7 @@
 /**
  * File containing the main intro class
  *
+ * @since   0.7.0 Dependency Injection Refactoring
  * @since   0.1.0
  * @package Eightshift_Libs\Core
  */
@@ -9,6 +10,8 @@
 namespace Eightshift_Libs\Core;
 
 use Eightshift_Libs\Exception;
+
+use \DI\ContainerBuilder;
 
 /**
  * The main start class.
@@ -62,12 +65,13 @@ abstract class Main implements Service {
   }
 
   /**
-   * Register the individual services of this theme/plugin.
+   * Register the individual services with optional dependency injection.
    *
    * @throws Exception\Invalid_Service If a service is not valid.
    *
    * @return void
    *
+   * @since 0.7.0 Dependency Injection Refactoring
    * @since 0.1.0
    */
   public function register_services() : void {
@@ -77,17 +81,104 @@ abstract class Main implements Service {
       return;
     }
 
-    $classes = $this->get_service_classes();
+    $this->services = $this->get_service_classes_with_di();
 
-    $this->services = array_map(
-      [ $this, 'instantiate_service' ],
-      $classes
-    );
     array_walk(
       $this->services,
-      function( Service $service ) {
-        $service->register();
+      function( $class ) {
+        if ( ! $class instanceof Registrable ) {
+          return;
+        }
+
+        $class->register();
       }
+    );
+  }
+
+  /**
+   * Return array of services with Dependency Injection parameters.
+   *
+   * @return array
+   *
+   * @since 0.7.0 Init
+   */
+  private function get_service_classes_with_di() : array {
+    $services = $this->get_service_classes_prepared_array();
+
+    $container = $this->get_di_container( $services );
+
+    return array_map(
+      function( $class ) use ( $container ) {
+        return $container->get( $class );
+      },
+      array_keys( $services )
+    );
+  }
+
+  /**
+   * Get services classes array and prepare it for dependency injection.
+   * Key should be a class name, and value should be an empty array or the dependencies of the class.
+   *
+   * @return array
+   *
+   * @since 0.7.0 Init.
+   */
+  private function get_service_classes_prepared_array() : array {
+    $output = [];
+    foreach ( $this->get_service_classes() as $class => $dependencies ) {
+      if ( gettype( $dependencies ) !== 'array' ) {
+        $output[ $dependencies ] = [];
+      } else {
+        $output[ $class ] = $dependencies;
+      }
+    }
+
+    return $output;
+  }
+
+  /**
+   * Implement PHP-DI.
+   * Build and return a DI container.
+   * Wire all the dependencies automatically, based on the provided array of class => dependencies from the get_di_items().
+   *
+   * @param array $services Array of service.
+   * @return object
+   *
+   * @since 0.7.0 Init.
+   */
+  private function get_di_container( array $services ) {
+    $builder = new ContainerBuilder();
+
+    $definitions = [];
+    foreach ( $services as  $service_key => $service_values ) {
+      if ( gettype( $service_values ) !== 'array' ) {
+        continue;
+      }
+
+      $definitions[ $service_key ] = \DI\create()->constructor( ...$this->get_di_dependencies( $service_values ) );
+    }
+
+    return $builder->addDefinitions( $definitions )->build();
+  }
+
+  /**
+   * Return prepared Dependency Injection objects.
+   * If you pass a class use PHP-DI to prepare if not just output it.
+   *
+   * @param array $dependencies Array of classes/parameters to push in constructor.
+   * @return array
+   *
+   * @since 0.7.0 Init
+   */
+  private function get_di_dependencies( array $dependencies ) : array {
+    return array_map(
+      function( $dependency ) {
+        if ( class_exists( $dependency ) ) {
+          return \DI\get( $dependency );
+        }
+        return $dependency;
+      },
+      $dependencies
     );
   }
 
@@ -108,7 +199,7 @@ abstract class Main implements Service {
     }
 
     $service = new $class();
-    if ( ! $service instanceof Service ) {
+    if ( ! $service instanceof Registrable ) {
       throw Exception\Invalid_Service::from_service( $service );
     }